+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.7.0
+### Added
+- Add `Serialize` instance for network constants.
+- Add `Serialize` instance for addresses that includes network constants.
+
+### Changed
+- Move functions related to addresses from `Script` to `Address` module.
+
 ## 0.6.1
 ### Added
 - Compatibility with latest GHC and base.